fix versions and refactor functio location
